+++ ledger_validation_result.js
@@ -1,3 +1,5 @@
+const {AssetProof} = require('./asset_proof');
+
 /**
  * To handle LedgerValidationResponse
  */
@@ -9,6 +11,17 @@ class LedgerValidationResult {
   constructor(code, proof) {
     this.code = code;
     this.proof = proof;
+  }
+
+  /**
+   * @param {LedgerValidationResponse} response
+   * @return {LedgerValidationResult}
+   */
+  static fromGRPCLedgerValidationResponse(response) {
+    return new LedgerValidationResult(
+        response.getStatusCode(),
+        AssetProof.fromGRPCAssetProof(response.getProof()),
+    );
   }
 
   /**
